Break up these single utils into utls and ext
where utils is static and ext is an extension, so utils would be named after the return type while ext is named after the receiver
