+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/InstantUtils.kt
@@ -2,6 +2,4 @@ package com.github.saikcaskey.pokertracker.domain.util
 
 import kotlinx.datetime.*
 
-fun nowAsInstant(): Instant {
-    return Clock.System.now()
-}
+fun nowAsInstant(): Instant = Clock.System.now()
