+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/LocalDateTimeUtils.kt
@@ -1,5 +1,6 @@
 package com.github.saikcaskey.pokertracker.domain.util
 
+import com.github.saikcaskey.pokertracker.domain.extensions.asLocalDateTime
 import kotlinx.datetime.*
 
 fun nowAsLocalDateTime(): LocalDateTime {
