+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/LocalDateUtils.kt
@@ -1,12 +1,12 @@
 package com.github.saikcaskey.pokertracker.domain.util
 
+import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalTime
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.atTime
 import kotlinx.datetime.toInstant
-import kotlinx.datetime.Instant
 
 fun LocalDate.atTimeInstant(parsedTime: LocalTime, timeZone: TimeZone? = null): Instant {
-    return atTime(parsedTime).toInstant(timeZone ?: TimeZone.currentSystemDefault())
+    return atTime(parsedTime).toInstant(timeZone ?: TimeZone.Companion.currentSystemDefault())
 }
