--- libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/asInstantOrNull.kt
@@ -1,7 +1,0 @@
-package com.github.saikcaskey.pokertracker.domain.util
-
-import kotlinx.datetime.*
-
-fun String?.asInstantOrNull(): Instant? {
-    return this?.let { runCatching(Instant::parse).getOrNull() }
-}
