--- libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/asLocalDateTime.kt
@@ -1,10 +1,0 @@
-package com.github.saikcaskey.pokertracker.domain.util
-
-import kotlinx.datetime.LocalDateTime
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
-import kotlinx.datetime.Instant
-
-fun Instant.asLocalDateTime(timeZone: TimeZone? = null): LocalDateTime {
-    return toLocalDateTime(timeZone ?: TimeZone.currentSystemDefault())
-}
